+++ src/pages/AboutGamePage/AboutGamePage.module.scss
@@ -1,0 +1,14 @@
+.root {
+  padding: 24px 0;
+  background: #f7f7f7;
+  width: 100%;
+}
+
+.header {
+  text-align: center;
+}
+
+.paragraph {
+  padding: 10px 0;
+  text-align: justify;
+}
