+++ src/pages/AboutGamePage/index.js
@@ -1,0 +1,3 @@
+import AboutGamePage from './AboutGamePage';
+
+export default AboutGamePage;
