+++ src/pages/CharactersPage/CharactersPage.js
@@ -1,0 +1,10 @@
+import s from './CharactersPage.module.scss';
+const CharactersPage = () => {
+  return (
+    <div className={s.root}>
+    
+    </div>
+  );
+}
+ 
+export default CharactersPage;
