+++ src/pages/CharactersPage/CharactersPage.module.scss
@@ -1,0 +1,1 @@
+.root {}
