+++ src/pages/CharactersPage/index.js
@@ -1,0 +1,3 @@
+import CharactersPage from './CharactersPage';
+
+export default CharactersPage;
