+++ src/pages/ContactsPage/ContactsPage.js
@@ -1,0 +1,10 @@
+import s from './ContactsPage.module.scss';
+const ContactsPage = () => {
+  return (
+    <div className={s.root}>
+    
+    </div>
+  );
+}
+ 
+export default ContactsPage;
