+++ src/pages/ContactsPage/index.js
@@ -1,0 +1,3 @@
+import ContactsPage from './ContactsPage';
+
+export default ContactsPage;
